Added hidden class, variables, functions, and dice image change.
Created const variables to be assigned with element classes. These elements are then used in functions to change their values. The variables: scores, currentScore, and activePlayer were created. An event listener was added to the btnRoll that would perform a function to randomize a number on the die and display it using an image source. Then a conditional statement was added if the dice rolled a 1 or not.