Disabled roll and hold button when a player has won
Disabled these functions by using a boolean. If the users are still playing, the boolean will allow a condition to use the functions that allow the dice to roll and the player to hold their score. If it's false, then these functions will not be accessible when roll or hold buttons are pressed.